# Trace optimizations