Following up from a mailing list discussion I took some measurements on where the time is spent on a simple Continue()+FunctionArguments() loop and made some optimizations taking the time spent on a single loop iteration from ~40ms to under 1ms.

The bulk of the improvement comes from skipping a time.Sleep in the linux implementation of singleStep, this sleep is needed to work around a quirk of the linux kernel but that particular quirk can not happen when we are stepping over a breakpoint, so it's unnecessary in that case.

The rest of the optimizations have a much smaller impact but cumulate to ~50% of the remaining time.

```
* proc: Added trace benchmark

Results:

BenchmarkTrace-4   	    5000	  36195899 ns/op

* proc/linux: faster single step implementation.

BenchmarkTrace-4   	    5000	   2093271 ns/op

* proc: Cache function debug_info entries to speed up variable lookup.

BenchmarkTrace-4   	    5000	   1864846 ns/op

* proc/variables: Optimize FunctionArguments by prefetching frame

BenchmarkTrace-4   	    5000	   1815795 ns/op

* proc/variables: optimized parseG

BenchmarkTrace-4   	   10000	    712767 ns/op
```